Fix kdump bitmap and page header handling
Use bitmap2, not the first bitmap, to know if a page is present.  The
first bitmap is only for memory holes, which we don't care about.  The
second bitmap is the one giving presence in the dump.

If a page is not present in the dump, then it's page header is not
written.  So account for this.

Signed-off-by: Corey Minyard <cminyard@mvista.com>