Store the elf phdrs in a btree for faster lookup
This required re-organizing how the phdrs are stored so that the
pointers to the data don't change with insertions and deletions.

So the data is still in an array, but an array of pointers to phdrs, and
an index is added to the phdr data so the btree can find it's way back
to the array index.

Signed-off-by: Corey Minyard <cminyard@mvista.com>