Reduce elfc tempfile usage
elfc was creating a tempfile per phdr.  On files with tens of thousands
of phdrs, you would get EMFILE.

Rework the code to have a single tempfile for all phdrs in a elf file
and use sections of the file to store data for specific phdrs.

Do the same for shdrs, since it shares the same architecture and some
code.

Signed-off-by: Corey Minyard <cminyard@mvista.com>